+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -837,4 +837,13 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
 
   @override
   String get addACommentDescription => "Permettre aux autres d'ajouter des commentaires";
+
+  @override
+  String get notifyChannelText => 'Notifier tout le monde dans ce canal';
+
+  @override
+  String get notifyHereText => 'Notifier tous les membres en ligne de ce canal';
+
+  @override
+  String notifyRoleText(String role) => 'Notifier tous les membres $role';
 }