+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ja.dart
@@ -813,4 +813,13 @@ class StreamChatLocalizationsJa extends GlobalStreamChatLocalizations {
 
   @override
   String get addACommentDescription => '他のユーザーにコメントの追加を許可';
+
+  @override
+  String get notifyChannelText => 'このチャンネルの全員に通知';
+
+  @override
+  String get notifyHereText => 'このチャンネルのオンラインメンバー全員に通知';
+
+  @override
+  String notifyRoleText(String role) => '$role 全員に通知';
 }